Restart Scripts
2 New Scripts that monitor your Muncoin Masternode.

blockMonitor.sh - monitors the block, if it detects that it is stuck, it kicks off a reindex of your node and flags for a restart command.

mnRestart.sh - watches for the blockMonitor to advise its kicked off a reindex, and watches the mn status to make sure its "insync" when it is, it tells the node to start.